fix free

+++ builtins/ft_export.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_export.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mnachit <mnachit@student.42.fr>            +#+  +:+       +#+        */
+/*   By: moer-ret <moer-ret@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/18 14:00:10 by monachit          #+#    #+#             */
-/*   Updated: 2024/06/03 15:04:36 by mnachit          ###   ########.fr       */
+/*   Updated: 2024/06/03 15:35:18 by moer-ret         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,7 +64,7 @@ char *check_value(char *value)
 {
     char *tmp;
     char *tmp2;
-    int i;
+    size_t i;
 
     i = 0;
     while (value[i] && value[i] != '=')
